+++ src/Facades/StrToken.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Fomvasss\LaravelStrTokens\Facades;
+
+use Fomvasss\LaravelStrTokens\StrTokenGenerator;
+use Illuminate\Support\Facades\Facade;
+
+class StrToken extends Facade
+{
+    public static function getFacadeAccessor()
+    {
+        return StrTokenGenerator::class;
+    }
+}