+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog Laravel Str Tokens
 
+## 0.2.1 - 2018-10-28
+
+- Added support `many` (relation) - always get first model
+- Added support taxonomy relation (prefix `tx`) (from package (fomvasss/laravel-taxonomy)[https://github.com/fomvasss/laravel-taxonomy])
+
 ## 0.0.1 - 2018-10-27
 
 - Initial commit